test/test_remote/requirements.txt
websocket_client
paho-mqtt==1.2